+++ src/main/java/com/practical_interview/project/controllers/dtos/TransactionRequest.java
@@ -2,9 +2,11 @@ package com.practical_interview.project.controllers.dtos;
 
 import com.practical_interview.project.persistence.entities.enums.TransactionTypeEnum;
 
+import java.math.BigDecimal;
+
 public record TransactionRequest(
         TransactionTypeEnum transactionType,
-        Long transactionAmount,
+        BigDecimal transactionAmount,
 
         String userId
 ) {
