+++ src/main/java/com/practical_interview/project/controllers/dtos/TransactionResponse.java
@@ -3,9 +3,11 @@ package com.practical_interview.project.controllers.dtos;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class TransactionResponse {
 
-    private Long accountBalance;
+    private BigDecimal accountBalance;
 }
