+++ src/main/java/com/practical_interview/project/controllers/dtos/TransferRequest.java
@@ -1,10 +1,12 @@
 package com.practical_interview.project.controllers.dtos;
 
+import java.math.BigDecimal;
+
 public record TransferRequest(
         String fromCustomerId,
         String toCustomerId,
 
-        Long transferAmount
+        BigDecimal transferAmount
 ) {
 
 }
